Build fixed (cast in webapi)

+++ src/server/webapi/datacoll.cpp
@@ -197,9 +197,9 @@ int H_DataCollectionHistory(Context *context)
       int pos = 1;
       DBBind(hStmt, pos++, DB_SQLTYPE_INTEGER, dci->getId());
       if (timeFrom != 0)
-         DBBind(hStmt, pos++, DB_SQLTYPE_INTEGER, timeFrom);
+         DBBind(hStmt, pos++, DB_SQLTYPE_INTEGER, static_cast<int>(timeFrom));
       if (timeTo != 0)
-         DBBind(hStmt, pos++, DB_SQLTYPE_INTEGER, timeTo);
+         DBBind(hStmt, pos++, DB_SQLTYPE_INTEGER, static_cast<int>(timeTo));
 
       DB_UNBUFFERED_RESULT hResult = DBSelectPreparedUnbuffered(hStmt);
       if (hResult != nullptr)
